requirements.txt
typer==0.4.0
torch==1.10.0
torchvision==0.10.0
torchaudio==0.9.0
gitpython==3.1.24
